test(client): move ethers integration tests
