+++ packages/client/tests/integration/helpers.ts
@@ -1,0 +1,19 @@
+import type {
+  AcceptedOrderResponse,
+  OrderResponse,
+} from '@polymarket/bindings/clob';
+import { expect } from 'vitest';
+
+export function expectAcceptedOrderResponse(
+  response: OrderResponse,
+): AcceptedOrderResponse {
+  expect(response.ok).toBe(true);
+
+  if (!response.ok) {
+    throw new Error(
+      `Expected accepted order response, received: ${response.code}`,
+    );
+  }
+
+  return response;
+}
